+++ 3week/e-commerce_3week/src/main/java/com/example/ecommerce_3week/dto/order/controller/OrderRequest.java
@@ -1,5 +1,6 @@
 package com.example.ecommerce_3week.dto.order.controller;
 
+import com.example.ecommerce_3week.dto.order.facade.OrderFacadeRequest;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -16,5 +17,11 @@ public class OrderRequest {
     public static class OrderItemRequest {
         private Long productId;
         private int quantity;
+    }
+
+    public static List<OrderFacadeRequest> from(OrderRequest request) {
+        return request.getItems().stream()
+                .map(i -> new OrderFacadeRequest(i.getProductId(), i.getQuantity()))
+                .toList();
     }
 }
